requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
regex
